Error handling when dll cannot be loaded
My AutoHotkey was set to run ahk files as 32-bit for some reason. The auto_oculus_touch.dll failed to load because of this (it seems to be 64-bit), but the script just kept happily running without any errors.

I added an error message to the script that alerts the user if the dll couldn't be loaded, and stops the script. I also added more descriptive error message to my particular problem.